[INLONG-11624][Agent] Modify code based on comments
